Bootstrap: fixed bugs in wget installation and openssl removal scripts

+++ bootstrap/os/slackware64/scripts/in-chroot/install-wget.sh
@@ -47,7 +47,9 @@ cat $INSTALL_SCRIPT.dist \
 ./$INSTALL_SCRIPT &&
 
 ### Create symlink for just in case
-ln -s /usr/local/bin/wget /usr/bin/wget
+if [ ! -f /usr/bin/wget ]; then
+    ln -s /usr/local/bin/wget /usr/bin/wget
+fi
